Handle notice only case
